// lib/faust-types.ts
export interface ShowcaseMemberFieldsType {
  members: ShowcaseMember[];
}

export interface ShowcaseMemberNode {
  id: string;
  title: string;
  showcaseMemberFieldsType: ShowcaseMemberFieldsType;
}

export interface ShowcaseMembersData {
  showcaseMembers: {
    nodes: ShowcaseMemberNode[];
  };
}